Move yezzey proxy proto IO util in separate package
